+++ meta_package_manager/managers/mas.py
@@ -41,10 +41,6 @@ class MAS(PackageManager):
     # /mas-cli/mas/commit/ca72ee42b1c5f482513b1d2fbf780b0bf3d9618b
     requirement = '>= 1.3.1'
 
-    def __init__(self):
-        super(MAS, self).__init__()
-        self.map = {}
-
     @cachedproperty
     def name(self):
         return "Mac AppStore"
@@ -65,11 +61,11 @@ class MAS(PackageManager):
         for application in output.split('\n'):
             if not application:
                 continue
-            _id, name, installed_version, latest_version = regexp.match(
-                application).groups()
-            self.map[name] = _id
+            package_id, package_name, installed_version, latest_version = \
+                regexp.match(application).groups()
             self.updates.append({
-                'name': name,
+                'id': package_id,
+                'name': package_name,
                 'latest_version': latest_version,
                 # Normalize unknown version. See: https://github.com/mas-cli
                 # /mas/commit/1859eaedf49f6a1ebefe8c8d71ec653732674341
@@ -77,11 +73,8 @@ class MAS(PackageManager):
                     installed_version if installed_version != 'unknown'
                     else None)})
 
-    def update_cli(self, package_name):
-        if package_name not in self.map:
-            return []
-        return [self.cli_path] + self.cli_args + [
-            'install', self.map[package_name]]
+    def update_cli(self, package_id):
+        return [self.cli_path] + self.cli_args + ['install', package_id]
 
     def update_all_cli(self):
         return [self.cli_path] + self.cli_args + ['upgrade']