+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-`1.12.0 (unreleased) <https://github.com/kdeldycke/meta-package-manager/compare/v1.11.0...develop>`_
+`1.12.0 (2016-12-03) <https://github.com/kdeldycke/meta-package-manager/compare/v1.11.0...v1.12.0>`_
 ----------------------------------------------------------------------------------------------------
 
 * Rename ``mpm update`` command to ``mpm upgrade``.